pyproject.toml
[tool.poetry]
name = "demotivatory"
version = "0.1.0"
description = ""
authors = ["Anatoly Raev <cralixraev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
discord = "^1.7.3"
async-cse = {git = "https://github.com/CralixRaev/async-cse-TheRealchr1s"}
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
